requirements.txt
Flask==2.0.2
numpy==1.22.2
opencv_python==4.5.3.56
Pillow==8.0.1
tensorflow==2.8.0
